263 add username validation to server

+++ routes/ideas.js
@@ -54,17 +54,28 @@ router.post('/', async (req, res) => {
 // update idea
 router.put('/:id', async (req, res) => {
     try {
-        const updatedIdea = await Idea.findByIdAndUpdate(
-            req.params.id,
-            {
-                $set: {
-                    text: req.body.text,
-                    tag: req.body.tag,
+        const idea = await Idea.findById(req.params.id);
+
+        // match the usernames
+        if (idea.username === req.body.username) {
+            const updatedIdea = await Idea.findByIdAndUpdate(
+                req.params.id,
+                {
+                    $set: {
+                        text: req.body.text,
+                        tag: req.body.tag,
+                    },
                 },
-            },
-            { new: true }
-        );
-        res.json({ success: true, data: updatedIdea });
+                { new: true }
+            );
+            return res.json({ success: true, data: updatedIdea });
+        }
+
+        // username do not math
+        res.status(403).json({
+            success: false,
+            error: 'you are not authorized to update this resource',
+        });
     } catch (error) {
         console.log(error);
         res.status(500).json({ success: false, error: 'something went wrong' });
@@ -74,14 +85,23 @@ router.put('/:id', async (req, res) => {
 // delete an idea
 router.delete('/:id', async (req, res) => {
     try {
-        await Idea.findByIdAndDelete(req.params.id);
-        res.json({ success: true, data: {} });
+        const idea = await Idea.findById(req.params.id);
+
+        // match the usernames
+        if (idea.username === req.body.username) {
+            await Idea.findByIdAndDelete(req.params.id);
+            return res.json({ success: true, data: {} });
+        }
+
+        // usernames do not match
+        res.status(403).json({
+            success: false,
+            error: 'you are not authorized to delete this resource',
+        });
     } catch (error) {
         console.log(error);
         res.status(500).json({ success: false, error: 'something went wrong' });
     }
-
-    res.json({ success: true, data: {} });
 });
 
 // always export the router
